Port to java-gitlab-api v4
Note that this commit will not build, because the java-gitlab-api dependency
in pom.xml is currently at 1.1.19, which does not have this API. A pre-built
version with that API is not available, yet, so we're using using a custom
built release from our internal repository.

There's no point in making this internal dependency public though, so it is
left out here.